Clear the job and the panel references from memory when both the query file and the viewer are closed

+++ packages/extension/src/runner.ts
@@ -1,53 +1,40 @@
-import { format as formatBytes } from "bytes";
 import {
-  createClient,
   createCSVFormatter,
   createFileOutput,
-  createFlat,
   createJSONFormatter,
   createJSONLinesFormatter,
   createLogOutput,
   createMarkdownFormatter,
   createTableFormatter,
   createViewerOutput,
-  DryRunJob,
   Formatter,
   Output,
-  RunInfo,
-  RunJob,
 } from "core";
-import { Results } from "core/src/types";
 import { createWriteStream } from "fs";
-import { readFile } from "fs/promises";
 import mkdirp from "mkdirp";
 import { basename, extname, join } from "path";
-import {
-  ExtensionContext,
-  Range,
-  TextDocument,
-  Uri,
-  WebviewPanel,
-  window,
-  workspace,
-} from "vscode";
+import { Range, TextDocument, workspace } from "vscode";
 import { OutputChannel, Result } from ".";
 import { Config } from "./config";
 import { ConfigManager } from "./configManager";
-import { ErrorMarker } from "./errorMarker";
+import { PanelManager } from "./panelManager";
+import { Renderer } from "./renderer";
+import { RunJobManager, RunJobResponse } from "./runJobManager";
 import { StatusManager } from "./statusManager";
 
 export type Runner = ReturnType<typeof createRunner>;
-export type DryRunner = ReturnType<typeof createDryRunner>;
 
 export function createRunner({
   outputChannel,
   outputManager,
+  panelManager,
   statusManager,
   runJobManager,
   renderer,
 }: {
   readonly outputChannel: OutputChannel;
   readonly outputManager: OutputManager;
+  readonly panelManager: PanelManager;
   readonly statusManager: StatusManager;
   readonly runJobManager: RunJobManager;
   readonly renderer: Renderer;
@@ -148,73 +135,21 @@ export function createRunner({
     },
 
     onDidCloseTextDocument({ document }: { readonly document: TextDocument }) {
-      outputManager.delete({ document });
-      runJobManager.delete({ document });
-    },
-
-    dispose() {
-      // do nothing
-    },
-  };
-}
-
-export function createDryRunner({
-  outputChannel,
-  configManager,
-  statusManager,
-  errorMarker,
-}: {
-  readonly outputChannel: OutputChannel;
-  readonly configManager: ConfigManager;
-  readonly statusManager: StatusManager;
-  readonly errorMarker: ErrorMarker;
-}) {
-  return {
-    async run({
-      document,
-      selection,
-    }: {
-      readonly document: TextDocument;
-      readonly selection?: Range;
-    }): Promise<Result> {
-      try {
-        outputChannel.appendLine(`Dry run`);
-        statusManager.loadProcessed({
-          document,
-        });
-
-        const config = configManager.get();
-        const client = await createClient(config);
-
-        let job!: DryRunJob;
-        try {
-          errorMarker.clear({ document });
-          job = await client.createDryRunJob({
-            query: getQueryText({ document, range: selection }),
-          });
-          errorMarker.clear({ document });
-        } catch (err) {
-          errorMarker.mark({ document, err, selection });
-          throw err;
-        }
-
-        outputChannel.appendLine(`Job ID: ${job.id}`);
-        const { totalBytesProcessed } = job.getInfo();
-        const bytes = formatBytes(totalBytesProcessed);
-        outputChannel.appendLine(`Result: ${bytes} estimated to be read`);
-
-        statusManager.succeedProcessed({
-          document,
-          processed: {
-            bytes,
-          },
-        });
-
-        return { jobId: job.id };
-      } catch (err) {
-        statusManager.errorProcessed({ document });
-        throw err;
+      if (panelManager.exists({ document })) {
+        return;
       }
+      runJobManager.delete({ document });
+      panelManager.delete({ document });
+    },
+
+    onDidDisposePanel({ document }: { readonly document: TextDocument }) {
+      if (
+        workspace.textDocuments.some((d) => d.fileName === document.fileName)
+      ) {
+        return;
+      }
+      runJobManager.delete({ document });
+      panelManager.delete({ document });
     },
 
     dispose() {
@@ -282,229 +217,6 @@ export function createOutputManager({
       }
     },
 
-    delete({ document }: { readonly document: TextDocument }) {
-      panelManager.delete({ document });
-    },
-
-    dispose() {
-      // do nothing
-    },
-  };
-}
-
-export type PanelManager = ReturnType<typeof createPanelManager>;
-
-export function createPanelManager({
-  ctx,
-}: {
-  readonly ctx: ExtensionContext;
-}) {
-  const map: Map<string, WebviewPanel> = new Map();
-
-  return {
-    async get({
-      document,
-    }: {
-      readonly document: TextDocument;
-    }): Promise<WebviewPanel> {
-      const p = map.get(document.fileName);
-      if (p) {
-        p.reveal(undefined, true);
-        return p;
-      }
-
-      const root = join(ctx.extensionPath, "out/viewer");
-      const base = Uri.file(root)
-        .with({
-          scheme: "vscode-resource",
-        })
-        .toString();
-      const html = (await readFile(join(root, "index.html"), "utf-8")).replace(
-        "<head>",
-        `<head><base href="${base}/" />`
-      );
-      const panel = window.createWebviewPanel(
-        `bigqueryRunner:${document.fileName}`,
-        basename(document.fileName),
-        { viewColumn: -2, preserveFocus: true },
-        {
-          enableScripts: true,
-          localResourceRoots: [Uri.file(root)],
-        }
-      );
-      map.set(document.fileName, panel);
-      panel.iconPath = Uri.file(
-        join(ctx.extensionPath, "out/assets/icon-small.png")
-      );
-      panel.webview.html = html;
-      panel.onDidDispose(() => {
-        console.log("onDidDispose:", document.fileName);
-        map.delete(document.fileName);
-      });
-      ctx.subscriptions.push(panel);
-      return panel;
-    },
-
-    delete({ document }: { readonly document: TextDocument }) {
-      map.delete(document.fileName);
-    },
-
-    dispose() {
-      map.clear();
-    },
-  };
-}
-
-export type RunJobManager = ReturnType<typeof createRunJobManager>;
-
-export type RunJobResponse = {
-  jobId: string;
-  results: Results;
-  info: RunInfo;
-};
-
-export function createRunJobManager({
-  configManager,
-  errorMarker,
-}: {
-  readonly configManager: ConfigManager;
-  readonly errorMarker: ErrorMarker;
-}) {
-  const map: Map<string, RunJob> = new Map();
-
-  return {
-    async rows({
-      document,
-      selection,
-    }: {
-      readonly document: TextDocument;
-      readonly selection?: Range;
-    }): Promise<RunJobResponse> {
-      const config = configManager.get();
-      const client = await createClient(config);
-      let job: RunJob | undefined;
-      try {
-        errorMarker.clear({ document });
-        job = await client.createRunJob({
-          query: getQueryText({ document, range: selection }),
-          maxResults: config.pagination.results,
-        });
-        errorMarker.clear({ document });
-      } catch (err) {
-        errorMarker.mark({ document, err, selection });
-        throw err;
-      }
-      if (!job) {
-        throw new Error(`no job`);
-      }
-      map.set(document.fileName, job);
-      return {
-        jobId: job.id,
-        results: await job.getRows(),
-        info: await job.getInfo(),
-      };
-    },
-
-    async prevRows({
-      document,
-    }: {
-      readonly document: TextDocument;
-    }): Promise<RunJobResponse> {
-      const job = map.get(document.fileName);
-      if (!job) {
-        throw new Error(`no job`);
-      }
-      return {
-        jobId: job.id,
-        results: await job.getPrevRows(),
-        info: await job.getInfo(),
-      };
-    },
-
-    async nextRows({
-      document,
-    }: {
-      readonly document: TextDocument;
-    }): Promise<RunJobResponse> {
-      const job = map.get(document.fileName);
-      if (!job) {
-        throw new Error(`no job`);
-      }
-      return {
-        jobId: job.id,
-        results: await job.getNextRows(),
-        info: await job.getInfo(),
-      };
-    },
-
-    delete({ document }: { readonly document: TextDocument }) {
-      return map.delete(document.fileName);
-    },
-
-    dispose() {
-      map.clear();
-    },
-  };
-}
-
-export type Renderer = ReturnType<typeof createRenderer>;
-
-export function createRenderer({
-  outputChannel,
-  statusManager,
-}: {
-  readonly outputChannel: OutputChannel;
-  readonly statusManager: StatusManager;
-}) {
-  return {
-    async render({
-      document,
-      output,
-      response: {
-        jobId,
-        results,
-        info: { query, schema, numRows },
-      },
-    }: {
-      readonly document: TextDocument;
-      readonly output: Output;
-      readonly response: RunJobResponse;
-    }) {
-      try {
-        statusManager.loadBilled({ document });
-
-        outputChannel.appendLine(`Result: ${results.structs.length} rows`);
-        const bytes = formatBytes(parseInt(query.totalBytesBilled, 10));
-        outputChannel.appendLine(
-          `Result: ${bytes} to be billed (cache: ${query.cacheHit})`
-        );
-
-        const flat = createFlat(schema.fields);
-        await output.writeHeads({ flat });
-        await output.writeRows({ ...results, numRows, flat });
-
-        // const bytesWritten = await output.bytesWritten();
-        // if (bytesWritten !== undefined) {
-        //   outputChannel.appendLine(
-        //     `Total bytes written: ${formatBytes(bytesWritten)}`
-        //   );
-        // }
-
-        statusManager.succeedBilled({
-          document,
-          billed: { bytes, cacheHit: query.cacheHit },
-        });
-      } catch (err) {
-        statusManager.errorBilled({ document });
-        // statusManager.hide();
-        if (jobId) {
-          throw new ErrorWithId(err, jobId);
-        } else {
-          throw err;
-        }
-      }
-    },
-
     dispose() {
       // do nothing
     },
@@ -538,7 +250,7 @@ function createFormatter({ config }: { config: Config }): Formatter {
   }
 }
 
-function getQueryText({
+export function getQueryText({
   document,
   range,
 }: {
